Move reaction actions to reactionsStore
