+++ src/store/messagesStore.js
@@ -40,8 +40,6 @@ import {
 	getMessageContext,
 	postNewMessage,
 	postRichObjectToConversation,
-	addReactionToMessage,
-	removeReactionFromMessage,
 } from '../services/messagesService.js'
 import { useChatExtrasStore } from '../stores/chatExtras.js'
 import { useGuestNameStore } from '../stores/guestName.js'
@@ -1338,73 +1336,6 @@ const actions = {
 		const response = await postNewMessage(message, { silent: false })
 		return response
 
-	},
-
-	/**
-	 * Adds a single reaction to a message for the current user.
-	 *
-	 * @param {*} context the context object
-	 * @param {*} param1 conversation token, message id and selected emoji (string)
-	 */
-	async addReactionToMessage(context, { token, messageId, selectedEmoji }) {
-		try {
-			context.commit('addReactionToMessage', {
-				token,
-				messageId,
-				reaction: selectedEmoji,
-			})
-			// The response return an array with the reaction details for this message
-			const response = await addReactionToMessage(token, messageId, selectedEmoji)
-
-			const reactionsStore = useReactionsStore()
-			reactionsStore.updateReactions({
-				token,
-				messageId,
-				reactionsDetails: response.data.ocs.data,
-			})
-		} catch (error) {
-			// Restore the previous state if the request fails
-			context.commit('removeReactionFromMessage', {
-				token,
-				messageId,
-				reaction: selectedEmoji,
-			})
-			showError(t('spreed', 'Failed to add reaction'))
-		}
-	},
-
-	/**
-	 * Removes a single reaction from a message for the current user.
-	 *
-	 * @param {*} context the context object
-	 * @param {*} param1 conversation token, message id and selected emoji (string)
-	 */
-	async removeReactionFromMessage(context, { token, messageId, selectedEmoji }) {
-		try {
-			context.commit('removeReactionFromMessage', {
-				token,
-				messageId,
-				reaction: selectedEmoji,
-			})
-			// The response return an array with the reaction details for this message
-			const response = await removeReactionFromMessage(token, messageId, selectedEmoji)
-
-			const reactionsStore = useReactionsStore()
-			reactionsStore.updateReactions({
-				token,
-				messageId,
-				reactionsDetails: response.data.ocs.data,
-			})
-		} catch (error) {
-			// Restore the previous state if the request fails
-			context.commit('addReactionToMessage', {
-				token,
-				messageId,
-				reaction: selectedEmoji,
-			})
-			console.error(error)
-			showError(t('spreed', 'Failed to remove reaction'))
-		}
 	},
 
 	async removeExpiredMessages(context, { token }) {
